refactor: separate IAgreement interface from schema
